メイン画面のデータ登録はReactのpostで行うため実装とテストの修正をした
