+++ TODO.md
@@ -2,11 +2,12 @@
 
 ## SPA化の前に最低限のフロントデザインを作る
 - 収支を登録したら画面遷移することなく欄に項目と金額が追加される
-    - まずは参考になりそうな資料を調査
-        - なんかいい方法ある？
-    - 入力してボタンを押した際にreact側でフォームから入力値を取得してaxiosでpostリクエスト(onclickでメソッド指定、メソッドの中にaxios) -> サーバー側に入力値を送信してDB登録 + 登録したデータをテーブルに追加して表示させる 
-        - ボタンを押した(post、DBに登録した)後にaxiosのgetを動かす(どうやって動かす？勝手に動きはしないよね。axiosでpostしたメソッドの中でaxios getのメソッド呼ぶ？) -> jsxのテーブルを更新する
+    - 入力してボタンを押した際にreact側でフォームから入力値を取得してaxiosでpostリクエスト(onclickでメソッド指定、メソッドの中にpostのaxios) -> サーバー側に入力値を送信してDB登録 + 登録したデータをテーブルに追加して表示させる 
+    - 登録ボタンを押した段階でReact側で入力値を取得してそのデータをテーブルに追加する
+        - get通信がない分通信のコストは下がる(処理スピードも上がる)
+        - リロードしたら普通にgetで取ったデータが入るので問題ない
 - こけているテストの修正・削除
+    - ControllerとかAPIに移したのはいらんくなる
 
 - メイン画面も定期収支画面と同じようなjsにする
